requirements.txt
pydantic
fastapi
uvicorn
anthropic
python-dotenv
sentence-transformers
faiss-cpu
numpy
aiofiles
